requirements.txt
Django==2.0.3
django-bootstrap4==0.0.6
django-registration==2.4.1
pytz==2018.3
